fixed ART repo checks errors
